OSIDB-4243: Allow using cookies for requesting token refresh (#1)
For browser-based applications where it's difficult to securely store
the long-lived refresh token, a secure, httpOnly, sameSite=strict cookie
is sent to the browser when authenticating.

A new view will use said cookie and return an access token if the cookie
is valid.

This unburdens the browser-based application from having to handle such
token.